Updated imports and moved logic
