+++ secureenclave/secureenclave.py
@@ -17,6 +17,7 @@ from .gpg import Gpg
 from .gpgagent import GpgAgent
 from .smartcard import SmartCard
 from .identities import IdentityManager
+from .key_handler import KeyHandler
 
 
 __author__ = 'Gonzalo Alvarez'
@@ -24,12 +25,6 @@ __program__ = 'SecureEnclave'
 
 __gpg_fetch_key__ = """admin
 fetch
-quit
-"""
-
-__gpg_trust_key__ = """trust
-5
-y
 quit
 """
 
@@ -49,6 +44,7 @@ class SecureEnclave(object):
         self.gpg_agent = GpgAgent(self.gpg)
         self.smartcard = SmartCard(self.gpg)
         self.identity_manager = IdentityManager(self.home)
+        self.key_handler = KeyHandler(self.gpg)
 
     def __enter__(self):
         """Enter context manager"""
@@ -68,84 +64,10 @@ class SecureEnclave(object):
         """Exit context manager"""
         self.gpg_agent.stop()
 
-    def import_key(self, filename):
-        gpg_cmd = '{} --import {}'.format(self.gpg.getbin(), filename)
-        invoke.run(gpg_cmd, env=self.gpg.getenv(), pty=True)
-
-    def new_key(self, new_key_uid, passphrase):
-        if not new_key_uid:
-            logger.error("New key UID not provided.")
-            return False
-
-        logger.info(f"Creating new GPG key for UID: {new_key_uid}")
-        gpg_cmd = '{} -q --batch --passphrase {} --quick-generate-key "{}" rsa4096 cert never'.format(self.gpg.getbin(), passphrase, new_key_uid)
-        result = invoke.run(gpg_cmd, env=self.gpg.getenv(), pty=True)
-        if result.exited != 0:  # type:ignore
-            logger.error("Could not create the master GPG key properly.")
-            logger.error(f"GPG command output: {result.stdout} {result.stderr}")
-            return False
-
-        created_key = None
-        for key_attempt in self.gpg.get_keys():
-            if new_key_uid in key_attempt.uid:
-                created_key = key_attempt
-                break
-
-        if not created_key:
-            logger.error(f"Failed to find the newly created GPG key with UID part: {new_key_uid}. Please check GPG manually.")
-            available_keys_uids = [k.uid for k in self.gpg.get_keys()]
-            logger.debug(f"Available key UIDs after creation attempt: {available_keys_uids}")
-            return False
-
-        logger.debug(f'New GPG key created with fingerprint {created_key.fingerprint}')
-        logger.info('New GPG key created. Proceeding to create its subkeys.')
-        for subkey_type in ['sign', 'encrypt', 'auth']:
-            logger.info(f"Creating {subkey_type} subkey...")
-            gpg_cmd = '{} -q --batch --pinentry-mode=loopback --passphrase {} --quick-add-key "{}" rsa4096 "{}" "2y"'.format(self.gpg.getbin(), passphrase, created_key.fingerprint, subkey_type)
-            result = invoke.run(gpg_cmd, env=self.gpg.getenv(), pty=True)
-            if result.exited != 0:  # type:ignore
-                logger.error(f'Could not create {subkey_type} subkey properly.')
-                logger.error(f"GPG command output: {result.stdout} {result.stderr}")
-                return False
-            logger.success(f'{subkey_type.capitalize()} subkey created successfully.')
-        logger.success('GPG Key creation completed, including all subkeys. Use "key list" to explore.')
-        return True
-
-    def del_key(self, public=True, secret=True):
-        keys = self.gpg.get_keys()
-        if not keys:
-            logger.info("No keys available to delete.")
-            return
-
-        selected = Bullet('Select which key to delete: ', keys).launch()  # type:ignore
-        if not selected:
-            logger.info("Key deletion cancelled.")
-            return
-
-        if secret:
-            gpg_cmd_secret = '{} -q --batch --delete-secret-key {}'.format(self.gpg.getbin(), selected.fingerprint)
-            result_secret = invoke.run(gpg_cmd_secret, env=self.gpg.getenv(), pty=True, hide=True)
-            if result_secret.ok:  # type: ignore
-                logger.success(f"Secret key for {selected.fingerprint} deleted successfully.")
-            else:
-                logger.error(f"Failed to delete secret key for {selected.fingerprint}.")
-                logger.debug(f"Output: {result_secret.stdout}")  # type: ignore
-                logger.debug(f"Error: {result_secret.stderr}")  # type: ignore
-        else:
-            logger.info(f"Skipping secret key deletion for {selected.fingerprint} as per --secret flag.")
-
-        if public:
-            logger.info(f"Attempting to delete public key for {selected.fingerprint}...")
-            gpg_cmd_public = '{} -q --batch --delete-key {}'.format(self.gpg.getbin(), selected.fingerprint)
-            result_public = invoke.run(gpg_cmd_public, env=self.gpg.getenv(), pty=True, hide=True)
-            if result_public.ok:  # type: ignore
-                logger.success(f"Public key for {selected.fingerprint} deleted successfully.")
-            else:
-                logger.error(f"Failed to delete public key for {selected.fingerprint}.")
-                logger.debug(f"Output: {result_public.stdout}")  # type: ignore
-                logger.debug(f"Error: {result_public.stderr}")  # type: ignore
-        else:
-            logger.info(f"Skipping public key deletion for {selected.fingerprint} as per --secret flag.")
+    def decrypt(self, input, output):
+        logger.debug('Decrypting with GPG')
+        gpg_cmd = '{} --quiet --armor --decrypt -o {} {}'.format(self.gpg.getbin(), output, input)
+        invoke.run(gpg_cmd, env=self.gpg.getenv(), hide=False, pty=True)
 
     def encrypt(self, input, output):
         key_list = self.gpg.get_keys()
@@ -155,26 +77,3 @@ class SecureEnclave(object):
             logger.debug(f'Encrypting with keyid [{key_id}]')
             gpg_cmd = '{} --quiet --armor --encrypt --recipient {} -o {} {}'.format(self.gpg.getbin(), key_id, output, input)
             invoke.run(gpg_cmd, env=self.gpg.getenv(), hide=False, pty=True)
-
-    def trust_keys(self):
-        key_list = self.gpg.get_keys()
-        logger.debug(len(key_list))
-        logger.debug(key_list)
-        trusted_count = 0
-        for key in key_list:
-            if key.trust == 'ultimate':
-                logger.debug('Key already trusted {key.uid}')
-            else:
-                trusted_count += 1
-                logger.info(f'Key UID: {key.uid}')
-                client = YesNo(f'Would you like to trust [{key.fingerprint}] ', default='n')
-                if client.launch():
-                    gpg_cmd = '{} --quiet --expert --batch --display-charset utf-8 --command-fd 0 --no-tty --edit-key {}'.format(self.gpg.getbin(), key.fingerprint)
-                    invoke.run(gpg_cmd, env=self.gpg.getenv(), hide=True, in_stream=StringIO(__gpg_trust_key__))
-        if trusted_count == 0:
-            logger.info('No untrusted keys to trust')
-
-    def decrypt(self, input, output):
-        logger.debug('Decrypting with GPG')
-        gpg_cmd = '{} --quiet --armor --decrypt -o {} {}'.format(self.gpg.getbin(), output, input)
-        invoke.run(gpg_cmd, env=self.gpg.getenv(), hide=False, pty=True)
